Implemented Relationship with EntityFramework

+++ Services/CharacterService/CharacterService.cs
@@ -10,7 +10,7 @@ using netCore3.Data;
 using netCore3.Dtos.Character;
 using netCore3.Models;
 
-namespace netCore3.Services.CharacterServices
+namespace netCore3.Services.CharacterService
 {
     public class CharacterService : ICharacterService
     {
@@ -27,7 +27,6 @@ namespace netCore3.Services.CharacterServices
             _httpContextAccessor = httpContextAccessor;
             _context = context;
             _mapper = mapper;
-
         }
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharcters()
         {
@@ -40,12 +39,19 @@ namespace netCore3.Services.CharacterServices
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
-            Character dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
-            serviceResponse.Data = _mapper.Map<GetCharacterDto>(_context.Characters.FirstOrDefault(c => c.Id == id));
+            Character dbCharacter = await _context.Characters
+            .Include(c => c.Weapon)
+            .Include(c => c.CharacterSkills)
+            .ThenInclude(cs => cs.Skill)
+            .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+            // serviceResponse.Data = _mapper.Map<GetCharacterDto>(_context.Characters.FirstOrDefault(c => c.Id == id));
+            serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
 
         private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
@@ -55,7 +61,7 @@ namespace netCore3.Services.CharacterServices
             // character.Id = characters.Max(c => c.Id) + 1;
             await _context.Characters.AddAsync(character);
             await _context.SaveChangesAsync();
-            serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+            serviceResponse.Data = (_context.Characters.Where(c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
             return serviceResponse;
         }
 
@@ -65,13 +71,21 @@ namespace netCore3.Services.CharacterServices
             try
             {
 
-                Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
-                character.Name = updateCharacter.Name;
-                character.Class = updateCharacter.Class;
-                character.Defence = updateCharacter.Defence;
-                character.Hitpoint = updateCharacter.Hitpoint;
-                character.intelligence = updateCharacter.intelligence;
-                character.Strength = updateCharacter.Strength;
+                Character character = await _context.Characters.Include(c => c.User).FirstOrDefaultAsync(c => c.Id == updateCharacter.Id);
+                if (character.User.Id == GetUserId())
+                {
+                    character.Name = updateCharacter.Name;
+                    character.Class = updateCharacter.Class;
+                    character.Defence = updateCharacter.Defence;
+                    character.Hitpoint = updateCharacter.Hitpoint;
+                    character.intelligence = updateCharacter.intelligence;
+                    character.Strength = updateCharacter.Strength;
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found";
+                }
 
                 _context.Characters.Update(character);
                 await _context.SaveChangesAsync();
@@ -93,11 +107,22 @@ namespace netCore3.Services.CharacterServices
             try
             {
 
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
+                Character character = await _context.Characters
+                .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
 
-                serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                if (character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
+                    serviceResponse.Data = (_context.Characters.Where(c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found";
+                }
+
             }
             catch (Exception ex)
             {
